api/Python/requirements.txt
bson==0.5.10
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
dnspython==2.1.0
Flask==2.0.2
Flask-Cors==3.0.10
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.4
PyJWT==2.3.0
pymongo==3.12.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
